Created a handle Submit componets for all handles
