+++ Context/Dashboard/Admin/AdminContext.jsx
@@ -3,9 +3,29 @@ import { createContext, useState } from 'react';
 export const AdminContext = createContext();
 
 export const AdminProvider = ({ children }) => {
-  const [state, setState] = useState({});
+  const [state, setState] = useState({
+    name: '',
+    error: '',
+    success: '',
+    buttonText: 'Create',
+    image: '',
+  });
+
+  const [content, setContent] = useState('');
+  const [imageUploadButtonName, setImageUploadButtonName] =
+    useState('Upload image');
+
   return (
-    <AdminContext.Provider value={{ state, setState }}>
+    <AdminContext.Provider
+      value={{
+        state,
+        setState,
+        content,
+        setContent,
+        imageUploadButtonName,
+        setImageUploadButtonName,
+      }}
+    >
       {children}
     </AdminContext.Provider>
   );
